fix : #139-vacation 휴가일수 컬럼 자료형 double형으로 변경

+++ lms-api/src/main/java/com/savelms/api/user/service/UserService.java
@@ -138,9 +138,8 @@ public class UserService {
                         .getTeam()
                         .getValue())
                     .vacation(u.getVacations().stream()
-                        .map(v ->
-                            v.getRemainingDays() - v.getUsedDays())
-                        .reduce(0L, Long::sum))
+                        .map(v -> v.getRemainingDays() - v.getUsedDays())
+                        .reduce(0.0, Double::sum))
                     .build())
             .collect(Collectors.toList());
         return ListResponse.<UserResponseDto>builder()
